refactor(admin): extract StatusBadge, RowActions, and table helpers from external-data.tsx
Pure moves (no behavior change): StatusBadge → external-data-status-badge.tsx,
RowActions → external-data-row-actions.tsx, SortableHeader/toggleSort/compareValues
+ sort types → external-data-table-helpers.tsx. external-data.tsx: 2383 → 2056 lines.
